% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mc.R
\name{choose_variable}
\alias{choose_variable}
\title{choose_variable}
\usage{
choose_variable(k, m)
}
\arguments{
\item{k}{the number of the first fixed variables}

\item{m}{the number of all variables}
}
\value{
The function value
}
\description{
helper function(choose variables function)
}
\examples{
\dontrun{
choose_variable(2,8)
}
}
