% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs}
\alias{gibbs}
\alias{gibbsC}
\title{gibbs sample by Rcpp}
\usage{
gibbsC(N)
}
\arguments{
\item{N}{the sample size}
}
\value{
a Numeric Matrix,sp stands for random sample of multivariate distribution.
}
\description{
Generate the gibbs sample by Rcpp,using the method of gibbs sampling.
}
\examples{
\dontrun{
res <- gibbsC(5000)
}
}
