% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mc.R
\name{TSLS_SAIC}
\alias{TSLS_SAIC}
\title{TSLS_SAIC}
\usage{
TSLS_SAIC(G, T, M, Mfixed)
}
\arguments{
\item{G}{the number of replications}

\item{T}{the number of observations}

\item{M}{the number of instrumental variables}

\item{Mfixed}{the number of first fixed instrumental variables}
}
\value{
The function value
}
\description{
helper function(a simple Monte Carlo simulation to obtain properties of Model Averaging estimators using SAIC based on TSLS.)
}
\examples{
\dontrun{
TSLS_SAIC(5000, 100, 10, 2)
}
}
